Remove unnecessary unboxing

+++ src/main/java/org/apache/commons/csv/CSVRecord.java
@@ -122,7 +122,7 @@ public final class CSVRecord implements Serializable, Iterable<String> {
                 headerMap.keySet()));
         }
         try {
-            return values[index.intValue()];
+            return values[index];
         } catch (final ArrayIndexOutOfBoundsException e) {
             throw new IllegalArgumentException(String.format(
                 "Index for header '%s' is %d but CSVRecord only has %d values!", name, index, values.length));
@@ -244,7 +244,7 @@ public final class CSVRecord implements Serializable, Iterable<String> {
      * @return whether a given column is mapped and has a value
      */
     public boolean isSet(final String name) {
-        return isMapped(name) && getHeaderMapRaw().get(name).intValue() < values.length;
+        return isMapped(name) && getHeaderMapRaw().get(name) < values.length;
     }
 
     /**
